Actor::preRender - 70%

+++ src/AAA.hpp
@@ -11,7 +11,7 @@ extern u8 Stage_actorFreezeFlag;
 
 //
 void func_02066fe8(void *, u32, u32);
-extern u32 data_0v000_020ca84c;
+extern u8 data_0v000_020ca84c;
 
 //
 void func_02044d94(void *, void *);
